+++ OniAccess/ModConfig.cs
@@ -9,5 +9,6 @@ namespace OniAccess {
 		public bool PipeShapeEarcons { get; set; } = false;
 		public bool PassabilityEarcons { get; set; } = false;
 		public bool AnnounceBiomeChanges { get; set; } = true;
+		public bool FlowSonification { get; set; } = false;
 	}
 }